Moved ia_correlations to more logical location
