+++ halotools/empirical_models/ia_models/__init__.py
@@ -1,4 +1,3 @@
 from .ia_model_components import *
 from .ia_strength_models import *
 from .watson_distribution import DimrothWatson
-from .ia_correlations import *
