+++ halotools/mock_observables/ia_correlations/ee_projected.py
@@ -1,6 +1,6 @@
 r"""
-Module containing the `~halotools.mock_observables.alignments.ed_projected` function used to
-calculate the ellipticity-direction (ED) projected correlation functon
+Module containing the `~halotools.mock_observables.alignments.ee_projected` function used to
+calculate the ellipticity-ellipticity (EE) projected correlation functon
 """
 
 from __future__ import absolute_import, division, print_function, unicode_literals
@@ -9,22 +9,22 @@ import numpy as np
 from math import pi
 
 from .alignment_helpers import process_projected_alignment_args
-from ....mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
+from ..mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
-from ....mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
-from ....mock_observables.pair_counters import positional_marked_npairs_xy_z, npairs_xy_z, marked_npairs_xy_z
+from ..pair_counters.mesh_helpers import _enforce_maximum_search_length
+from ..pair_counters import marked_npairs_xy_z, npairs_xy_z, marked_npairs_xy_z
 
-__all__ = ['ed_projected`']
+__all__ = ['ee_projected']
 __author__ = ['Duncan Campbell']
 
 
 np.seterr(divide='ignore', invalid='ignore')  # ignore divide by zero in e.g. DD/RR
 
 
-def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None, weights2=None,
+def ee_projected(sample1, orientations1, sample2, orientations2, rp_bins, pi_max, weights1=None, weights2=None,
         period=None, num_threads=1, approx_cell1_size=None, approx_cell2_size=None):
     r"""
-    Calculate the ellipticity-direction projected correlation function (ED), :math:`\omega(r_p)`.
+    Calculate the projected ellipticity-ellipticity projected correlation function (EE), :math:`\eta(r_p)`.
 
     Parameters
     ----------
@@ -42,6 +42,10 @@ def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None
     sample2 : array_like, optional
         Npts2 x 3 array containing 3-D positions of points.
 
+    orientations2 : array_like
+        Npts1 x 2 numpy array containing orientation vectors for each point in ``sample2``.
+        these will be normalized if not already.
+
     rp_bins : array_like
         array of boundaries defining the radial bins perpendicular to the LOS in which
         pairs are counted.
@@ -50,7 +54,7 @@ def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None
     pi_max : float
         maximum LOS distance defining the projection integral length-scale in the z-dimension.
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
-
+    
     weights1 : array_like, optional
         Npts1 array of weghts.  If this parameter is not specified, it is set to numpy.ones(Npts1).
 
@@ -89,20 +93,19 @@ def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None
     Returns
     -------
     correlation_function : numpy.array
-        *len(rp_bins)-1* length array containing the correlation function :math:`\omega(r_p)`
+        *len(rbins)-1* length array containing the correlation function :math:`\eta(r_p)`
         computed in each of the bins defined by input ``rp_bins``.
 
     Notes
     -----
-    The ellipticity-direction projected correlation function is defined as:
+    The ellipticity-ellipticity correlation function is defined as:
 
     .. math::
-        \omega = \frac{\sum_{i \neq j}w_iw_j|\hat{e}_i \cdot \hat{r}_{ij}|^2}{\sum_{i \neq j} w_iw_j} - \frac{1}{2}
+        \eta = \frac{\sum_{i \neq j}w_iw_j|\hat{e}_i \cdot \hat{e}_j|^2}{\sum_{i \neq j}} - \frac{1}{2}
 
-    where e.g. :math:`\hat{e}_i` is the orientation of the :math:`i`-th galaxy, and
-    :math:`\hat{r}_{ij}` is the normalized vector in the direction of the :math:`j`-th galaxy
-    from the :math:`i`-th galaxy. :math:`w_i` and :math:`w_j` are the weights associated with
-    the :math:`i`-th and :math:`j`-th galaxy. The weights default to 1 if not set.
+    where e.g. :math:`\hat{e}_i` is the orientation of the :math:`i`-th galaxy.
+    :math:`w_i` and :math:`w_j` are the weights associated with the :math:`i`-th 
+    and :math:`j`-th galaxy. The weights default to 1 if not set.
 
     Example
     --------
@@ -130,25 +133,25 @@ def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None
 
     >>> random_orientations = np.random.random((Npts,2))
 
-    We can the calculate the auto-ED correlation between these points:
+    We can the calculate the auto-EE correlation between these points:
 
     >>> rp_bins = np.logspace(-1,1,10)
     >>> pi_max = 0.2
-    >>> result = ed_projected(sample1, random_orientations, sample1, rp_bins, pi_max, period=Lbox)
+    >>> result = ee_projected(sample1, random_orientations, sample1, random_orientations, rp_bins, pi_max, period=Lbox)
 
     """
 
     # process arguments
     alignment_args = (sample1, orientations1, None, weights1,
-                      sample2, None, None, weights2,
+                      sample2, orientations2, None, weights2,
                       None, None, None, None)
     dum = 0.0  # dummy variable to store arguments not needed for this function
     sample1, orientations1, dum, weights1,\
-    sample2, dum, dum, weights2,\
+    sample2, orientations2, dum, weights2,\
     dum, dum, dum, dum = process_projected_alignment_args(*alignment_args)
 
     function_args = (sample1, rp_bins, pi_max, sample2, period, num_threads)
-    sample1, rp_bins, pi_bins, sample2, period, num_threads, PBCs = _ed_projected_process_args(*function_args)
+    sample1, rp_bins, pi_bins, sample2, period, num_threads, PBCs = _ee_projected_process_args(*function_args)
 
     # How many points are there (for normalization purposes)?
     N1 = len(sample1)
@@ -158,15 +161,17 @@ def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None
     marks1[:,0] = weights1
     marks1[:,1] = orientations1[:,0]
     marks1[:,2] = orientations1[:,1]
-    marks2 = weights2
-    marked_counts, counts = positional_marked_npairs_xy_z(sample1, sample2, rp_bins, pi_bins,
-                                period=period, weights1=marks1, weights2=marks2,
-                                weight_func_id=4, num_threads=num_threads,
-                                approx_cell1_size=approx_cell1_size,
-                                approx_cell2_size=approx_cell2_size)
-    marked_counts = np.diff(np.diff(marked_counts, axis=0),axis=1)
-    
-    # if no weights, use fast un-weihgted pair counter
+    marks2 = np.zeros((N2, 3))
+    marks2[:,0] = weights2
+    marks2[:,1] = orientations2[:,0]
+    marks2[:,2] = orientations2[:,1]
+    marked_counts = marked_npairs_xy_z(sample1, sample2, rp_bins, pi_bins,
+                                     period=period, weights1=marks1, weights2=marks2,
+                                     weight_func_id=15, num_threads=num_threads,
+                                     approx_cell1_size=approx_cell1_size, approx_cell2_size=approx_cell2_size)
+    marked_counts = np.diff(np.diff(marked_counts, axis=0), axis=1)
+
+    # if no weights, use fast un-weighted pair counter
     if np.all(weights1==1.0) & np.all(weights2==1.0):
         counts = npairs_xy_z(sample1, sample2, rp_bins, pi_bins,
                        period=period, num_threads=num_threads,
@@ -178,22 +183,18 @@ def ed_projected(sample1, orientations1, sample2, rp_bins, pi_max, weights1=None
                        period=period, num_threads=num_threads,
                        approx_cell1_size=approx_cell1_size,
                        approx_cell2_size=approx_cell2_size)
-    counts = np.diff(np.diff(counts, axis=0),axis=1)
+    counts = np.diff(np.diff(counts, axis=0), axis=1)
 
     return marked_counts/counts - 1.0/2.0
 
 
-def _ed_projected_process_args(sample1, rp_bins, pi_max, sample2, period, num_threads):
+def _ee_projected_process_args(sample1, rp_bins, pi_max, sample2, period, num_threads):
     r"""
     Private method to do bounds-checking on the arguments passed to
-    `~halotools.mock_observables.alignments.ed_3d`.
+    `~halotools.mock_observables.alignments.ee_projected`.
     """
 
     sample1 = enforce_sample_has_correct_shape(sample1)
-    sample2 = enforce_sample_has_correct_shape(sample2)
-    
-    N1 = len(sample1)
-    N2 = len(sample2)
 
     rp_bins = get_separation_bins_array(rp_bins)
     rp_max = np.amax(rp_bins)
