+++ halotools/mock_observables/ia_correlations/gi_minus_projected.py
@@ -1,33 +1,41 @@
 r"""
-Module containing the `~halotools.mock_observables.alignments.ii_plus_3d` function used to
-calculate the intrinsic ellipticity-ellipticity (II) correlation
+Module containing the `~halotools.mock_observables.alignments.gi_minus_projected` function used to
+calculate the projected gravitational shear-intrinsic ellipticity (GI) correlation
 """
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
-from math import pi, gamma
+from math import pi
 
-from .alignment_helpers import process_3d_alignment_args
-from ....mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
+from .alignment_helpers import process_projected_alignment_args
+from ..mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
-from ....mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
-from ....mock_observables.pair_counters import positional_marked_npairs_3d, marked_npairs_3d
+from ..pair_counters.mesh_helpers import _enforce_maximum_search_length
+from ..pair_counters import positional_marked_npairs_xy_z, marked_npairs_xy_z
 
-__all__ = ['ii_plus_3d']
+__all__ = ['gi_minus_projected']
 __author__ = ['Duncan Campbell']
 
 
 np.seterr(divide='ignore', invalid='ignore')  # ignore divide by zero in e.g. DD/RR
 
 
-def ii_plus_3d(sample1, orientations1, ellipticities1, sample2, orientations2, ellipticities2,
-            rbins, randoms1=None, randoms2=None, weights1=None, weights2=None,
-            ran_weights1=None, ran_weights2=None, estimator='Natural',
+def gi_minus_projected(sample1, orientations1, ellipticities1, sample2, rp_bins, pi_max,
+            randoms1=None, randoms2=None, weights1=None, weights2=None,
+            ran_weights1=None, ran_weights2=None, estimator='Landy-Szalay',
             period=None, num_threads=1, approx_cell1_size=None, approx_cell2_size=None):
     r"""
-    Calculate the intrinsic ellipticity-ellipticity correlation function (II),
-    :math:`\xi_{++}(r)`.  See the 'Notes' section for details of this calculation.
+    Calculate the projected gravitational shear-intrinsic ellipticity correlation function (GI),
+    :math:`w_{g-}(r_p)`, where :math:`r_p` is the separation perpendicular to the line-of-sight (LOS)
+    between two galaxies.  See the 'Notes' section for details of this calculation.
+
+    The first two dimensions define the plane for perpendicular distances.  The third
+    dimension is used for parallel distances, i.e. x,y positions are on the plane of the
+    sky, and z is the redshift coordinate. This is the 'distant observer' approximation.
+
+    Note in particular that the `~halotools.mock_observables.alignments.gi_minus_projected` function does not
+    accept angular coordinates for the input ``sample1`` or ``sample2``.
 
     Parameters
     ----------
@@ -40,26 +48,22 @@ def ii_plus_3d(sample1, orientations1, ellipticities1, sample2, orientations2, e
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
 
     orientations1 : array_like
-        Npts1 x 3 numpy array containing projected orientation vectors for each point in ``sample1``.
+        Npts1 x 2 numpy array containing projected orientation vectors for each point in ``sample1``.
         these will be normalized if not already.
 
     ellipticities1: array_like
         Npts1 x 1 numpy array containing ellipticities for each point in ``sample1``.
 
     sample2 : array_like, optional
-        Npts2 x 3 array containing 3-D positions of points with associated
-        orientations and ellipticities.
+        Npts2 x 3 array containing 3-D positions of points.
 
-    orientations2 : array_like
-        Npts1 x 3 numpy array containing projected orientation vectors for each point in ``sample2``.
-        these will be normalized if not already.
-
-    ellipticities2: array_like
-        Npts1 x 1 numpy array containing ellipticities for each point in ``sample2``.
-
-    rbins : array_like
-        array of boundaries defining the radial bins in which
+    rp_bins : array_like
+        array of boundaries defining the radial bins perpendicular to the LOS in which
         pairs are counted.
+        Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
+
+    pi_max : float
+        maximum LOS distance defining the projection integral length-scale in the z-dimension.
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
 
     randoms1 : array_like, optional
@@ -121,36 +125,44 @@ def ii_plus_3d(sample1, orientations1, ellipticities1, sample2, orientations2, e
     Returns
     -------
     correlation_function : numpy.array
-        *len(rbins)-1* length array containing the correlation function :math:`w_{g+}(r)`
-        computed in each of the bins defined by input ``rbins``.
+        *len(rp_bins)-1* length array containing the correlation function :math:`w_{g+}(r_p)`
+        computed in each of the bins defined by input ``rp_bins``.
 
     Notes
     -----
 
-    The II-correlation function is calculated as:
+    The projected GI-correlation function is calculated as:
 
     .. math::
-        \xi_{++}(r) = \frac{S_{+}S_{+}}{R_sR_s}
+        w_{g-}(r_p) = 2 \int_0^{\pi_{\rm max}} \xi_{g-}(r_p, \pi) \mathrm{d}\pi
 
-    where
+    If the Landy-Szalay estimator is indicated, the correlation function is estimated as:
 
     .. math::
-        S_{+}S_{+} = \sum_{i \neq j} w_jw_i e_{+}(j|i)e_{+}(i|j)
+        \xi_{g-}(r_p, \pi) = \frac{S_{-}D-S_{-}R}{R_sR}
+
+     where
+
+    .. math::
+        S_{-}D = \sum_{i \neq j} w_jw_i e_{-}(j|i)
 
     :math:`w_j` and :math:`w_j` are weights.  Weights are set to 1 for all galaxies by default.
     The alingment of the :math:`j`-th galaxy relative to the direction to the :math:`i`-th galaxy is given by:
 
     .. math::
-        e_{+}(j|i) = e_j\cos(2\phi)
+        e_{-}(j|i) = e_j\sin(2\phi)
 
     where :math:`e_j` is the ellipticity of the :math:`j`-th galaxy.  :math:`\phi` is the angle between the
-    orientation vector, :math:`\vec{o}_j`, and the direction between the :math:`j`-th
-    and :math:`i`-th galaxy, :math:`\vec{r}_{i,j}`.
+    orientation vector, :math:`\vec{o}_j`, and the projected direction between the :math:`j`-th
+    and :math:`i`-th galaxy, :math:`\vec{r}_{p i,j}`.
 
     .. math::
-        \cos(\phi) = \vec{o}_j \cdot \vec{r}_{i,j}
+        \cos(\phi) = \vec{o}_j \cdot \vec{r}_{p i,j}
 
-    :math:`R_sR_s` are random pair counts,
+    :math:`S_{-}R` is analgous to :math:`S_{-}D` but instead is computed
+    with respect to a "random" catalog of galaxies.  :math:`R_sR` are random pair counts,
+    where :math:`R_s` corresponds to the shapes sample, i.e. the sample with orienttions
+    and ellipticies, ``sample1``, and R correspoinds to ``sample2``.
 
     Examples
     --------
@@ -176,28 +188,29 @@ def ii_plus_3d(sample1, orientations1, ellipticities1, sample2, orientations2, e
 
     We then create a set of random orientation vectors and ellipticities for each point
 
-    >>> random_orientations = np.random.random((Npts,3))
+    >>> random_orientations = np.random.random((Npts,2))
     >>> random_ellipticities = np.random.random(Npts)
 
     We can the calculate the projected auto-GI correlation between these points:
 
-    >>> rbins = np.logspace(-1,1,10)
-    >>> w = ii_plus_3d(sample1, random_orientations, random_ellipticities, sample1, random_orientations, random_ellipticities, rbins, period=Lbox)
+    >>> rp_bins = np.logspace(-1,1,10)
+    >>> pi_max = 0.25
+    >>> w = gi_minus_projected(sample1, random_orientations, random_ellipticities, sample1, rp_bins, pi_max, period=Lbox)
 
     """
 
     # process arguments
     alignment_args = (sample1, orientations1, ellipticities1, weights1,
-                      sample2, orientations2, ellipticities2, weights2,
+                      sample2, None, None, weights2,
                       randoms1, ran_weights1, randoms2, ran_weights2)
     sample1, orientations1, ellipticities1, weights1, sample2,\
         orientations2, ellipticities2, weights2, randoms1, ran_weights1,\
-        randoms2, ran_weights2 = process_3d_alignment_args(*alignment_args)
+        randoms2, ran_weights2 = process_projected_alignment_args(*alignment_args)
 
-    function_args = (sample1, rbins, sample2, randoms1, randoms2,
+    function_args = (sample1, rp_bins, pi_max, sample2, randoms1, randoms2,
         period, num_threads, approx_cell1_size, approx_cell2_size)
-    sample1, rbins, sample2, randoms1, randoms2,\
-        period, num_threads, PBCs, no_randoms = _ii_plus_3d_process_args(*function_args)
+    sample1, rp_bins, pi_bins, sample2, randoms1, randoms2,\
+        period, num_threads, PBCs, no_randoms = _gi_minus_projected_process_args(*function_args)
 
     # How many points are there (for normalization purposes)?
     N1 = len(sample1)
@@ -211,99 +224,106 @@ def ii_plus_3d(sample1, orientations1, ellipticities1, sample2, orientations2, e
 
     #define merk vectors to use in pair counting
     # sample 1
-    marks1 = np.ones((N1, 4))
+    marks1 = np.ones((N1, 3))
     marks1[:, 0] = ellipticities1 * weights1
     marks1[:, 1] = orientations1[:, 0]
     marks1[:, 2] = orientations1[:, 1]
-    marks1[:, 3] = orientations1[:, 2]
     # sample 2
-    marks2 = np.ones((N2, 4))
-    marks2[:, 0] = ellipticities2 * weights2
-    marks2[:, 1] = orientations2[:, 0]
-    marks2[:, 2] = orientations2[:, 1]
-    marks2[:, 3] = orientations2[:, 2]
+    marks2 = weights2
     # randoms 1
-    ran_marks1 = np.ones((NR1, 4))
+    ran_marks1 = np.ones((NR1, 3))
     ran_marks1[:, 0] = ran_weights1
     ran_marks1[:, 1] = 0  # dummy
     ran_marks1[:, 2] = 0  # dummy
-    ran_marks1[:, 3] = 0  # dummy
     # randoms 2
-    ran_marks2 = np.ones((NR2, 4))
-    ran_marks2[:, 0] = ran_weights2
-    ran_marks2[:, 1] = 0  # dummy
-    ran_marks2[:, 2] = 0  # dummy
-    ran_marks2[:, 3] = 0  # dummy
+    ran_marks2 = ran_weights2
 
-    do_SS, do_RR = II_estimator_requirements(estimator)
+    # define pi bins
+    pi_bins = np.array([0.0, pi_max])
+
+    do_SD, do_SR, do_RR = GI_estimator_requirements(estimator)
 
     # count marked pairs
-    if do_SS:
-        SS = marked_pair_counts(sample1, sample2,  marks1,  marks2,
-                                rbins, period, num_threads,
+    if do_SD:
+        SD = marked_pair_counts(sample1, sample2,  marks1,  marks2,
+                                rp_bins, pi_bins, period, num_threads,
                                 approx_cell1_size, approx_cell2_size)
     else:
-        SS = None
+        SD = None
+
+    # count marked random pairs
+    if do_SR:
+        if no_randoms:
+            SR = 0.0
+        else:
+            SR = marked_pair_counts(sample1, randoms2, marks1, ran_marks2,
+                                    rp_bins, pi_bins, period, num_threads,
+                                    approx_cell1_size, approx_cell2_size)
+    else:
+        SR = None
 
     # count random pairs
     if do_RR:
         RR = random_counts(randoms1, randoms2, ran_weights1, ran_weights2,
-                           rbins, N1, N2, no_randoms, period, PBCs,
+                           rp_bins, pi_bins, N1, N2, no_randoms, period, PBCs,
                            num_threads, approx_cell1_size, approx_cell2_size)
     else:
         RR = None
 
-    result = II_estimator(SS, RR, N1, N2, NR1, NR2, estimator)
+    result = GI_estimator(SD, SR, RR, N1, N2, NR1, NR2, estimator)
 
-    return result  # factor of 2pi_max accounts for integration
+    return result*2.0*pi_max  # factor of 2pi_max accounts for integration
 
 
-def II_estimator(SS, RR, N1, N2, NR1, NR2, estimator='Natural'):
+def GI_estimator(SD, SR, RR, N1, N2, NR1, NR2, estimator='Landy-Szalay'):
     r"""
     apply the supplied GI estimator to calculate the correlation function.
     """
-    if estimator == 'Natural':
+
+    if estimator == 'Landy-Szalay':
         factor = (NR1*NR2)/(N1*N2)
-        return factor*(SS/RR)
+        return factor*(SD-SR)/RR
     else:
         msg = ('The estimator provided is not supported.')
         raise ValueError(msg)
 
 
-def II_estimator_requirements(estimator):
+def GI_estimator_requirements(estimator):
     r"""
     Return the requirments for the supplied GI estimator.
     """
-
+    do_SD = False
+    do_SR = False
     do_RR = False
-    do_SS = False
 
-    if estimator == 'Natural':
-        do_SS = True
+    if estimator == 'Landy-Szalay':
+        do_SD = True
+        do_SR = True
         do_RR = True
-        return do_SS, do_RR
+        return do_SD, do_SR, do_RR
     else:
         msg = ('The estimator provided is not supported.')
         raise ValueError(msg)
 
 
-def marked_pair_counts(sample1, sample2, weights1, weights2, rbins, period,
+def marked_pair_counts(sample1, sample2, weights1, weights2, rp_bins, pi_bins, period,
         num_threads, approx_cell1_size, approx_cell2_size):
     r"""
     Count marked pairs.
     """
 
-    weight_func_id = 5
-    SS = positional_marked_npairs_3d(sample1, sample2, rbins, period=period,
+    weight_func_id = 3
+    SD = positional_marked_npairs_xy_z(sample1, sample2, rp_bins, pi_bins, period=period,
         weights1=weights1, weights2=weights2, weight_func_id=weight_func_id,
         num_threads=num_threads, approx_cell1_size=approx_cell1_size,
         approx_cell2_size=approx_cell1_size)[0]
-    SS = np.diff(SS, axis=0)
+    SD = np.diff(np.diff(SD, axis=0), axis=1)
+    SD = SD.flatten()
 
-    return SS
+    return SD
 
 
-def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rbins,
+def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rp_bins, pi_bins,
                   N1, N2, no_randoms, period,
                   PBCs, num_threads, approx_cell1_size, approx_cell2_size):
     r"""
@@ -311,12 +331,13 @@ def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rbins,
     """
 
     if no_randoms is False:
-        RR = marked_npairs_3d(randoms1, randoms2, rbins,
+        RR = marked_npairs_xy_z(randoms1, randoms2, rp_bins, pi_bins,
                 period=period, num_threads=num_threads, weight_func_id=1,
                 weights1=ran_weights1, weights2=ran_weights2,
                 approx_cell1_size=approx_cell1_size,
                 approx_cell2_size=approx_cell2_size)
-        RR = np.diff(RR, axis=0)
+        RR = np.diff(np.diff(RR, axis=0), axis=1)
+        RR = RR.flatten()
 
         return RR
     else:
@@ -326,30 +347,29 @@ def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rbins,
         NR2 = N2
 
         # do volume calculations
-        v = nball_volume(rbins)
-        dv = np.diff(v, axis=0)
+        v = cylinder_volume(rp_bins, 2.0*pi_bins)
+        dv = np.diff(np.diff(v, axis=0), axis=1)
         global_volume = period.prod()
 
-        # calculate the random-random pairs.
+        # calculate the expected random-random pairs.
         rhor = (NR1*NR2)/global_volume
         RR = (dv*rhor)
 
         return RR.flatten()
 
 
-def nball_volume(R, k=3):
+def cylinder_volume(R, h):
+    r"""
+    Calculate the volume of a cylinder(s), used for the analytical randoms.
     """
-    Calculate the volume of a n-shpere.
-    This is used for the analytical randoms.
-    """
-    return (np.pi**(k/2.0)/gamma(k/2.0+1.0))*R**k
+    return pi*np.outer(R**2.0, h)
 
 
-def _ii_plus_3d_process_args(sample1, rbins, sample2, randoms1, randoms2,
+def _gi_minus_projected_process_args(sample1, rp_bins, pi_max, sample2, randoms1, randoms2,
         period, num_threads, approx_cell1_size, approx_cell2_size):
     r"""
     Private method to do bounds-checking on the arguments passed to
-    `~halotools.mock_observables.alignments.alignments.ii_plus_projected`.
+    `~halotools.mock_observables.alignments.gi_minus_projected`.
     """
     sample1 = enforce_sample_has_correct_shape(sample1)
 
@@ -376,12 +396,15 @@ def _ii_plus_3d_process_args(sample1, rbins, sample2, randoms1, randoms2,
     else:
         no_randoms = False
 
-    rbins = get_separation_bins_array(rbins)
-    rmax = np.amax(rbins)
+    pi_max = float(pi_max)
+    pi_bins = np.array([0.0, pi_max])
+
+    rp_bins = get_separation_bins_array(rp_bins)
+    rp_max = np.amax(rp_bins)
 
     period, PBCs = get_period(period)
 
-    _enforce_maximum_search_length([rmax, rmax, rmax], period)
+    _enforce_maximum_search_length([rp_max, rp_max, pi_max], period)
 
     if (randoms1 is None) & (PBCs is False):
         msg = "If no PBCs are specified, both randoms must be provided.\n"
@@ -389,6 +412,4 @@ def _ii_plus_3d_process_args(sample1, rbins, sample2, randoms1, randoms2,
 
     num_threads = get_num_threads(num_threads)
 
-    return sample1, rbins, sample2, randoms1, randoms2, period, num_threads, PBCs, no_randoms
-
-
+    return sample1, rp_bins, pi_bins, sample2, randoms1, randoms2, period, num_threads, PBCs, no_randoms
