+++ halotools/mock_observables/ia_correlations/tests/test_ed_3d.py
