+++ halotools/mock_observables/ia_correlations/tests/test_ed_projected.py
