+++ halotools/mock_observables/ia_correlations/tests/test_ee_3d.py
