+++ halotools/mock_observables/ia_correlations/tests/test_gi_minus_3d.py
