+++ halotools/mock_observables/ia_correlations/tests/test_gi_minus_projected.py
