+++ halotools/mock_observables/ia_correlations/tests/test_gi_plus_projected.py
