+++ halotools/mock_observables/ia_correlations/tests/test_ii_minus_projected.py
