+++ halotools/mock_observables/ia_correlations/tests/test_ii_plus_3d.py
